requirements.txt
fastapi
uvicorn
pydantic
instanseg
openslide-python
pillow
numpy
